requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.9.6
gunicorn==19.1.1
itsdangerous==0.24
wsgiref==0.1.2
